Moving imports from bitbucket to github
