+++ .travis.yml
@@ -1,5 +1,3 @@
-sudo: required
-
 services:
   - docker
 
@@ -18,7 +16,7 @@ env:
 after_success:
   - CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build main.go
   - docker login -e $DOCKER_EMAIL -u $DOCKER_USER -p $DOCKER_PASS
-  - export REPO=sebestblog/travis-demo
+  - export REPO=bign8/pipelines
   - export TAG=`if [ "$TRAVIS_BRANCH" == "master" ]; then echo "latest"; else echo $TRAVIS_BRANCH ; fi`
   - docker build -f Dockerfile -t $REPO:$COMMIT .
   - docker tag $REPO:$COMMIT $REPO:$TAG
