+++ attic/graph.go
@@ -1,4 +1,4 @@
-package pipelines
+package attic
 
 //
 // import "fmt"
